requirements.txt
discord.py
apscheduler
Flask
gunicorn
python-dotenv
aiosqlite